@@ -5,13 +5,8 @@ import {
   RouterProvider,
 } from 'react-router-dom'
 import './App.css'
-import AdminPage from './componentes/admin-page'
-import LoginForm from './componentes/auth/login-form'
-import CommandPage from './componentes/command/command'
-import CommandList from './componentes/command/command-list'
-import { CommandProvider } from './componentes/command/command-provider'
-import { CommandsProvider } from './componentes/command/commands-provider'
 import RootLayout from './componentes/root-layout'
+import LoginForm from './componentes/auth/login-form'
 import CreateDealPage from './componentes/create-deal-page'
 
 function App() {
@@ -20,24 +15,13 @@ function App() {
       <Route path='/'>
         <Route path='p' element={<RootLayout />}>
           <Route path='login' element={<LoginForm />} />
-          <Route path='manage' element={<AdminPage />} />
-          <Route path='commands' element={<CreateDealPage />} />
-          <Route
-            path='command'
-            element={
-              <CommandProvider>
-                <CommandPage />
-              </CommandProvider>
-            }
-          />
+          <Route path='transaction' element={<CreateDealPage />} />
         </Route>
       </Route>,
     ),
   )
   return (
-    <CommandsProvider>
       <RouterProvider router={router} />
-    </CommandsProvider>
   )
 }
 
