@@ -1,17 +1,35 @@
-import { AxiosInstance } from 'axios'
-
 export interface Token {
   access_token: TokenData
   token_type: string
 }
 
 export interface TokenData {
-  sub: string
-  scopes: string
+  sub: TokenSub
+  username: string
+}
+export interface TokenSub {
+  username: string
+  role: string
 }
 export interface Auth {
   isAutonticated: boolean
   user: string
-  scope: number
+  scope: string
   isLoading: boolean
+}
+
+export enum Role {
+  ADMIN = 'admin',
+  USER = 'user',
+}
+export interface LoginUser {
+  username: string
+  password: string
+}
+
+export interface RegisterUser {
+  username: string
+  password: string
+  balance: number
+  role: Role
 }
