@@ -1,16 +1,15 @@
+/* eslint-disable @typescript-eslint/no-empty-function */
 import jwtDecode from 'jwt-decode'
 import { createContext, useContext, useEffect, useState } from 'react'
 import { deepEqual } from '../../utils'
 import { useApi } from '../api/api-provider'
 import Provider from '../provider-model'
-import { Token } from './models'
+import { Token } from '../auth/models'
 
-const TokenContext = createContext<[Token | undefined, Function]>([
+const TokenContext = createContext<[Token | undefined, unknown]>([
   undefined,
   () => {},
 ])
-export const KEY =
-  '09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7'
 export const TokenProvider: React.FC<Provider> = ({ children }) => {
   const api = useApi()
   const [token, setToken] = useState<Token | undefined>(undefined)
@@ -48,7 +47,7 @@ const getTokenFromStorage = () => {
   return {}
 }
 
-const refreshToken = (state_token: Token | undefined, setToken: Function) => {
+const refreshToken = (state_token: Token | undefined, setToken: (token: object)=> void) => {
   const storage_token = getTokenFromStorage()
   if (
     !deepEqual(storage_token, state_token) &&
@@ -58,7 +57,7 @@ const refreshToken = (state_token: Token | undefined, setToken: Function) => {
   }
 }
 
-export const useToken = (): [Token | undefined, Function] => {
+export const useToken = (): [Token | undefined, unknown] => {
   const [state_token, setToken] = useContext(TokenContext)
   return [
     state_token,
