@@ -1,4 +1,4 @@
-export const deepEqual = (obj1: any, obj2: any) => {
+export const deepEqual = (obj1: object, obj2: object) => {
   // Check if both objects are of the same type
   if (typeof obj1 !== typeof obj2) {
     return false
